encryptioncheck.ps1
Ioan Popovici with the SCCM-Zone.com built a script that checks Bitlocker status on a workstation.  Great tool, but at my current customer, there's a mix of both Bitlocker and McAfee encryption.  I've updated his file to also include a check for McAfee Encryption, along with the status of the drive.